requirements.txt
py-cord==2.0.0b7
PyYAML==6.0
